Jail bookings schema, data faker, other fixes
In addition to a simple jail bookings schema, this includes a simple data faker with builtin matching.

Details about the faker:

The matching is done somewhat stupidly in that it can create matches within a file and then overwrite the matched rows when it creates matches between that file and another file. Files matched later will have fewer overwritten matches. This shouldn't be a problem for larger files (where the probability of getting the same row in two random draws is small) and is probably good enough for now.

It allows you to specify on the command line how many and what types of files you want to create, and how many matches you would like to make.

- Add jail bookings schema [Resolves #55]
- Add faker script [Resolves #83]
- Update HMIS and Jail schemas with lowercase/correct names and schema v3
- Rename and modify fake bookings file to include hmis and also making a few people match
- Include output of faker script
- Modify webapp to use lowercase id names [Resolves #56]
- Change merge output to pipe [Resolves #82]
- Add unit tests that ensures all schemas validate
- Make provider types (soon to be event types) less ambiguous
- Pin faker to make sure it has the right generators