If merged, this commit will make the following changes to the faker:
- allow you to sepcify from the command line how many and what types of files you want to create
- allow you to specify from the command line how many matches you would like to make
- enable creating matches within a file

The matching is done somewhat stupidly in that it can create matches within a file and then overwrite the matched rows when it creates matches between that file and another file. Files matched later will have fewer overwritten matches. This shouldn't be a problem for larger files (where the probability of getting the same row in two random draws is small) and is probably good enough for now.