Acceptance Testing [Resolves #132]
This commit adds acceptance testing to the uploader using nightwatch as a controller for a Selenium server. Basic flow tests are written for things like login, logout, successful and unsuccessful uploading attempts. Nightwatch has a feature called 'page objects' which is a way to reuse selectors and commands for a specific page. They are utilized heavily here.

There are also a couple of bugfixes here, illuminated by this testing.

Instructions for running these tests is in the README

- Add nightwatch package and config JSON file, using chrome for testing and two workers.
- Add login page object and flow test that covers login and logout
- Add homepage and upload page objects, and flow tests for HMIS stay and jail bookings upload
- Add simultaneous upload tests with separate counties, including a creation script for their users
- Fix jail bookings schema to no longer use booleans [Resolves #131]
- Fix logout link in header to bypass react-router and use the server [Resolves #130]
- Add instructions to README