Task Notification [Resolves #200 #202 #213]  (#231)
The task notification timeline shows the current matching tasks, tasks in queue and whole completed history. Refreshing the page will update the current matching task runtime. Once the current task is done, it will appear at the top of the completed task history and the task in the queue will be pushed to the current task.

Create a notification timeline of current task, task in queue and whole completed task history.
Remove "back to home button" during validation.