Make jail contacts and cumulative duration by booking [Resolves #280]
The jail booking schema allows input to specify location moves within a single jail booking. These show up as multiple rows, that all share the same jail_entry and jail_exit dates. To avoid double-counting bookings during aggregation, we need to ensure that we use the durations of the bookings and use those when needed.

In this case, both the table results query and the jail duration query needed to be modified.