Make uploader usable without s3 [Resolves #320]
Since much of the code was already using smart_open, the replacement here that uses boto3 has the same semantics as smart_open: an open function that switches between filesystem and s3 based on scheme.

To showcase this, there is a new endpoint test that does the while flow (uploading and merging) without mocking s3 at all to prove that it works.

- Introduce open_sesame to switch between s3 and filesystem
- Change references to 's3' in code to 'storage'
- Add bookings merge test that uses filesystem storage