Sort results with NULLs last [Resolves #328]:
This commit changes how results are sorted on the results page so that
regardless of what sort order a user requests for a given column, NULL
values always appear at the end of the list. It also updates the
documentation to specify this and to view NULLs for service-specific
columns.