Remove quotes from the example .env file
In the example .env file, where no default or example string is given for an environment variable, a pair of empty quotes is given. This may lead users to insert the replacement string _inside_ the quotes. This commit suggests replacing those quotes with the word _example_ to potentially avoid this issue.